requirements.txt
fastapi
uvicorn
google-cloud-speech
google-cloud-texttospeech
google-cloud-aiplatform
google-api-python-client
python-multipart
playsound==1.2.2